+++ sphinx/history.rst
@@ -10,6 +10,7 @@ History
 - Hide NotGeoreferencedWarning warning when subdatasets are present using open_rasterio (issue #65)
 - Add support for loading in 1D variables in `xarray.open_rasterio()` (issue #43)
 - Load in netCDF metadata on the variable level (pull #73)
+- Add rioxarray.merge module (issue #46)
 
 0.0.17
 ------